add support for cross-compiling
A new destination type, cross-compile, is introduced which indicates that the
destination for the package should support cross-compiling.  Currently, only the
exndbam format repoistory supports this functionality.

The exndbam repository is enhanced to support two new keys:
    - cross_compile_host
    - tool_prefix

The cross_compile_host key identifies the destination as a cross-compile
repository.  The tool_prefix key identifies the tool prefix used for the objcopy
and strip tools during the strip phase.  These are now passed as part of the
parameters to various actions within the e repository.

A new behaviour modifier, uncrossable, is also added to indicate that the
package (e.g. accounts) cannot be cross-compiled.

exheres-0 now includes exhost which adds interfaces for exheres to query
information about cross-compilation configuration.

build_functions.bash has been updated to follow the new directory structure as
specified by the multiarch specification.

Signed-off-by: Saleem Abdulrasool <compnerd@compnerd.org>