introduce the concept of a split_debug_location
This is intended to help support parallel installation of cross-compiled
packages.  A (package) installed repository (e.g. not accounts) can specify the
split_debug_location key to indicate that the debug information should be stored
in an alternate path.  For preserving backwards compatibility, the key is not
made mandatory.  Omission of the key defaults the path to the traditional FHS
path of /usr/lib/debug.